+++ src/desktop/platform.c
@@ -200,12 +200,12 @@ void br_plotter_init_specifics_platform(br_plotter_t* br, int width, int height)
 
 static void br_glfw_on_scroll(GLFWwindow* window, double x, double y) {
   (void)window;
-  stl_br->mouse.scroll = BR_VEC2(x, y);
+  stl_br->mouse.scroll = BR_VEC2((float)x, (float)y);
 }
 
 static void br_glfw_on_mouse_move(struct GLFWwindow * window, double x, double y) {
   (void)window;
-  stl_br->mouse.pos = BR_VEC2(x, y);
+  stl_br->mouse.pos = BR_VEC2((float)x, (float)y);
 }
 
 static void br_glfw_on_mouse_button(struct GLFWwindow* window, int button, int action, int mods) {
